# Library search slowing down a lot after putting space while typing in search bar