#### Purpose

Search is too slow for phrases which contains spaces. This PR tries to fix it.

#### Modifications

There was investigated that main reason of slowness is using of Regular Expressions. It is proposed to replace regualar expression by manual implementation. Because of used regular expression manual implementation looks not so complicated.

#### Performance check

All presented data counted in milliseconds. Were done 3 passes. Measured just search algorithm. Not the time needed for visual population of search results.

Letter entered|Old impl. average|New impl. average
--------------|-----------------|-----------------
d|	1.000|3.667
o|	0.667|1.000
u|	1.000|1.333
b|	1.000|1.000
l|	1.000|1.333
e|	1.000|1.000
 |	0.667|1.000
s|	523.697|1.333
l|	582.367|1.333
i|	604.368|1.333
d|	593.034|1.000
e|	591.034|1.000
r|	588.034|1.000

Current computer specification: Intel i7, 8GB RAM.

By the way as I noticed for regular expressions on i5 `slider` times are near 930 ms.

Let me know if more search phrases should be measured.

#### Search content
Here is below a couple snapshots of content for old implementation (left) and new implementation (right).
![capture](https://cloud.githubusercontent.com/assets/8158551/6105024/a214db30-b05b-11e4-915d-8c0930f83382.PNG)

Below found a case were results doesn't match. Which of sides is correct?
![capture](https://cloud.githubusercontent.com/assets/8158551/6105076/0a0b07c8-b05c-11e4-8c55-ac4dd2b705d6.PNG)

For next letter results are the same.
![capture](https://cloud.githubusercontent.com/assets/8158551/6105105/58d58928-b05c-11e4-8260-63a068eab7f6.PNG)

![capture](https://cloud.githubusercontent.com/assets/8158551/6105207/6c0156d4-b05d-11e4-92be-20cee2b6a8f0.PNG)

![capture](https://cloud.githubusercontent.com/assets/8158551/6105184/12eabaf4-b05d-11e4-84e6-560df33391c1.PNG)

#### Additional references

[MAGN-6018](http://adsk-oss.myjetbrains.com/youtrack/issue/MAGN-6018) Library search slowing down a lot after putting space while typing in search bar.

#### Unit tests
![capture](https://cloud.githubusercontent.com/assets/8158551/6105329/bf0c5c24-b05e-11e4-8b00-f03f147e844b.PNG)

#### Reviewers

@pboyer, @Benglin, please take a look.