@vmoyseenko 2015-02-09
@pboyer, @Benglin, please, take a look. We have difference in search results. You can find it on screenshot for `double s`. How should I handle it?