@Benglin 2015-02-09
Thanks for the fix, @vmoyseenko! This surely is exciting, look at that constant speed search! :)

As for the different result list when you searched for `double`, I think none of the two lists is *right*. Do you think we should not search into the argument list? Maybe @elayabharath can advise what's the desirable behavior here, but I think searching the argument string is a little overkill.