@pboyer 2015-02-09
Hi @vmoyseenko, 

This is a clear improvement.  I think there's opportunities to improve this algorithm a bit (e.g. search by characters and ignoring white space) and we may use it as the only search algorithm, but for now this fixes a critical performance issue.  

I differ with @Benglin  on argument lists.  I thin it's fine as it is.

I'm going to merge this in as it is.  If we want to make any of the changes @Benglin has made, we can tackle them in subsequent PR's.  

Thanks!